📜 assingment/assingment_o3/pyproject.toml
[project]
name = "assingment-03"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.6.2",
    "google-generativeai>=0.8.5",
    "openai-agents>=0.2.2",
    "python-dotenv>=1.1.1",
]
